Move approx() to its own module
Move (and rename) the function to make it easier to test.
