+++ lib/engine_util.js
@@ -14,6 +14,7 @@ const vm = require('vm');
 const A = require('async');
 const jsonpath = require('JSONPath');
 const cheerio = require('cheerio');
+const jitter = require('./jitter').jitter;
 
 let xmlCapture;
 try {
@@ -31,31 +32,6 @@ module.exports = {
   evil: evil
 };
 
-function approx(sApprox) {
-  if (!sApprox) {
-    return sApprox;
-  }
-
-  if (typeof sApprox !== 'string') {
-    return sApprox;
-  }
-
-  if (sApprox.indexOf(':') < 0) {
-    return sApprox;
-  }
-
-  let inputs = sApprox.split(':');
-  let nb = parseInt(inputs[0], 10);
-  let approxPercent = parseInt(inputs[1], 10);
-
-  let approx = approxPercent;
-  if (inputs[1].indexOf('%') >= 0) {
-    approx = nb * approxPercent / 100;
-  }
-
-  return Math.max(0, nb - approx + Math.random() * 2 * approx);
-}
-
 function createThink(requestSpec, opts) {
   opts = opts || {};
 
@@ -64,7 +40,7 @@ function createThink(requestSpec, opts) {
   let f = function(context, callback) {
     let thinktime = parseFloat(template(thinkspec, context)) * 1000;
     if (requestSpec.jitter || opts.jitter) {
-      thinktime = approx(`${thinktime}:${requestSpec.jitter || opts.jitter}`);
+      thinktime = jitter(`${thinktime}:${requestSpec.jitter || opts.jitter}`);
     }
     debug('think %s, %s, %s -> %s', requestSpec.think, requestSpec.jitter, opts.jitter, thinktime);
     setTimeout(function() {
